% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes_disposal}
\alias{codes_disposal}
\title{Cross-walk table for disposal codes.}
\format{
Tibble with rows for disposal codes and columns for definitions.
}
\usage{
codes_disposal
}
\description{
Cross-walk table with disposal codes and definitions.
}
\examples{
data(codes_disposal)
codes_disposal
}
\seealso{
Other data: 
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}}
}
\concept{data}
\keyword{datasets}
