% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes_gear}
\alias{codes_gear}
\title{Cross-walk table for gear codes.}
\format{
Tibble with rows for gear codes and columns for definitions.
}
\usage{
codes_gear
}
\description{
Cross-walk table with gear codes and definitions.
}
\examples{
data(codes_gear)
codes_gear
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}}
}
\concept{data}
\keyword{datasets}
