% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes_group}
\alias{codes_group}
\title{Cross-walk table for group codes.}
\format{
Tibble with rows for group codes and columns for definitions.
}
\usage{
codes_group
}
\description{
Cross-walk table with group codes and definitions.
}
\examples{
data(codes_group)
codes_group
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}}
}
\concept{data}
\keyword{datasets}
