% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes_period}
\alias{codes_period}
\title{Cross-walk table for period codes.}
\format{
Tibble with rows for period codes and columns for definitions.
}
\usage{
codes_period
}
\description{
Cross-walk table with period codes and definitions.
}
\examples{
data(codes_period)
codes_period
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}}
}
\concept{data}
\keyword{datasets}
