% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes_source}
\alias{codes_source}
\title{Cross-walk table for source codes.}
\format{
Tibble with rows for source codes and columns for definitions.
}
\usage{
codes_source
}
\description{
Cross-walk table with source codes and definitions.
}
\examples{
data(codes_source)
codes_source
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{conv_factors}},
\code{\link{database_info}},
\code{\link{unbalanced_sampling}},
\code{\link{undefined_sections}}
}
\concept{data}
\keyword{datasets}
