% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_factors}
\alias{conv_factors}
\title{Conversion factors.}
\format{
List with conversion factors.
}
\usage{
conv_factors
}
\description{
Conversion factors: think metric.
}
\examples{
data(conv_factors)
conv_factors
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}}
}
\concept{data}
\keyword{datasets}
