% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_info}
\alias{database_info}
\title{Database information.}
\format{
List with database information: databaseconnection and information
for tables: location, biosamples, catch, widths, surface spawn, macrocystis
spawn, understory spawn, and all spawn.
}
\usage{
database_info
}
\description{
Database information: connection, schemas, tables, and column names.
}
\examples{
data(database_info)
database_info
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}}
}
\concept{data}
\keyword{datasets}
