% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_bio}
\alias{load_bio}
\title{Load biological data.}
\usage{
load_bio(
  db_info = database_info,
  groups = codes_group,
  year_start = SpawnIndex::pars$years$assess,
  rep_year_start = 2014,
  quiet = FALSE
)
}
\arguments{
\item{db_info}{List. Herring database connection information from
\code{\link{database_info}}.}

\item{groups}{Tibble Spatial groups. Default is from
\code{\link{codes_group}}.}

\item{year_start}{Integer. First year to include in time series. Default is
from \code{SpawnIndex::pars$years$assess}.}

\item{rep_year_start}{Year indicating the start of representative samples. Default 2014.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Tibble with biological data (e.g., length, weight, age, sex)
including spatial and temporal information.
}
\description{
Load and wrangle biological data from the database.
}
\examples{
library(SpawnIndex)
library(DBI)
data(pars)
data(codes_group)
data(database_info)
bio_raw <- load_bio(db_info = database_info)
}
\seealso{
Other loaders: 
\code{\link{load_catch}()},
\code{\link{load_spawn}()}
}
\concept{loaders}
