% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_catch}
\alias{load_catch}
\title{Load catch data.}
\usage{
load_catch(
  db_info = database_info,
  groups = codes_group,
  conversions = conv_factors,
  year_start = SpawnIndex::pars$years$assess,
  quiet = FALSE
)
}
\arguments{
\item{db_info}{List. Herring database connection information from
\code{\link{database_info}}.}

\item{groups}{Tibble Spatial groups. Default is from
\code{\link{codes_group}}.}

\item{conversions}{List Conversion factors. Default is from
\code{\link{conv_factors}}.}

\item{year_start}{Integer. First year to include in time series. Default is
from \code{SpawnIndex::pars$years$assess}.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Tibble with catch data including spatial and temporal information.
Catch is in tonnes except for spawn on kelp (SOK) landings in pounds.
}
\description{
Load and wrangle catch data from the database.
}
\examples{
library(SpawnIndex)
library(DBI)
data(pars)
data(codes_group)
data(conv_factors)
data(database_info)
catch_raw <- load_catch()
}
\seealso{
Other loaders: 
\code{\link{load_bio}()},
\code{\link{load_spawn}()}
}
\concept{loaders}
