% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{map_base}
\alias{map_base}
\title{Base map for region.}
\usage{
map_base(shape_land, shape_sar, map_buffer = 25000)
}
\arguments{
\item{shape_land}{Simple feature object with BC coastline.}

\item{shape_sar}{Simple feature object with stock assessment region in
question.}

\item{map_buffer}{Numeric. Buffer around map in meters. Default 25,000.}

\item{french}{Logical. Translate to French. Default FALSE.}
}
\value{
Map of the BC coast with Pacific Herring stock assessment regions.
}
\description{
Base map for region.
}
\examples{
2 + 2
}
\seealso{
Other maps: 
\code{\link{map_bc}()}
}
\concept{maps}
