% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{map_bc}
\alias{map_bc}
\title{Map of BC.}
\usage{
map_bc(shape_land, shape_sars, map_buffer = 25000, french = FALSE)
}
\arguments{
\item{shape_land}{Simple feature object with BC coastline.}

\item{shape_sars}{Simple feature object with stock assessment regions.}

\item{map_buffer}{Numeric. Buffer around map in meters. Default 25,000.}

\item{french}{Logical. Translate to French. Default FALSE.}
}
\value{
Map of the BC coast with Pacific Herring stock assessment regions.
}
\description{
Map of BC with Pacific Herring stock assessment regions.
}
\examples{
require(SpawnIndex)
require(rnaturalearth)
data(sections)
data(database_info)
areas <- load_area_data(
  reg = "All", db = database_info$herring_con, where = database_info$area_loc
)
polys <- load_sections(sections = sections, areas = areas)
bc_coast <- ne_countries(
  scale = "large", returnclass = "sf",
  country = c("Canada", "United States of America")
)
map_bc(shape_land = bc_coast, shape_sars = polys$regions)
}
\seealso{
Other maps: 
\code{\link{map_base}()}
}
\concept{maps}
