% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mean_na_roll}
\alias{mean_na_roll}
\title{Calculate rolling mean.}
\usage{
mean_na_roll(x, n = 5)
}
\arguments{
\item{x}{Numeric. Values to mean.}

\item{n}{Numeric. Number of prior values in rolling window (maximum). Default
is 5.}
}
\value{
Numeric rolling mean.
}
\description{
Calculate rolling mean of previous values.
}
\examples{
vec <- c(1, 2, NA, 4, 5, 6, NA, NA, NA, 10)
vec_no_na <- mean_na_roll(x = vec)
}
\seealso{
Other utility: 
\code{\link{plus_group}()},
\code{\link{season_to_year}()}
}
\concept{utility}
