% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mean_na_weight}
\alias{mean_na_weight}
\title{Calculate weighted mean.}
\usage{
mean_na_weight(x, w, omit_na = TRUE)
}
\arguments{
\item{x}{Numeric. Values to mean.}

\item{w}{Numeric. Weights.}

\item{omit_na}{Logical. Keep or remove NAs. Default is TRUE.}
}
\value{
Numeric weighted mean.
}
\description{
Calculate weighted mean if there are non-NA values, return NA if all values
are NA. This does not fail when groups have no values to summarise.
}
\seealso{
Other utility: 
\code{\link{mean_na}()},
\code{\link{mean_na_roll}()},
\code{\link{season_to_year}()},
\code{\link{sum_na}()}
}
\concept{utility}
