% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plus_group}
\alias{plus_group}
\title{Wrangle age into plus group.}
\usage{
plus_group(x, max_age = 10)
}
\arguments{
\item{x}{Numeric. Ages to wrangle.}

\item{max_age}{Numeric. Maximum age; combine older into plus group.
Default 10.}
}
\value{
Numeric age.
}
\description{
Combine ages above oldest age into a plus group.
}
\seealso{
Other utility: 
\code{\link{mean_na_roll}()},
\code{\link{season_to_year}()}
}
\concept{utility}
