% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{report_table}
\alias{report_table}
\title{Default table.}
\usage{
report_table(dat, cap = "Table caption.")
}
\arguments{
\item{dat}{Tibble. Data for the table.}

\item{cap}{Character. Table caption. Default is "Table caption."}
}
\value{
kable of dat for RMarkdown.
}
\description{
Default data summary report table.
}
\concept{tables}
