% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rolling_mean_na}
\alias{rolling_mean_na}
\title{Calculate rolling mean.}
\usage{
rolling_mean_na(x, n = 5)
}
\arguments{
\item{x}{Numeric. Values to mean.}

\item{n}{Numeric. Number of values in rolling window (maximum). Default is 5.}
}
\value{
Numberic rolling mean.
}
\description{
Calculate rolling mean of previous values.
}
\seealso{
Other utility: 
\code{\link{mean_na}()},
\code{\link{season_to_year}()},
\code{\link{sum_na}()},
\code{\link{weighted_mean_na}()}
}
\concept{utility}
