% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{season_to_year}
\alias{season_to_year}
\title{Transform season to year.}
\usage{
season_to_year(dat)
}
\arguments{
\item{dat}{Numeric. Seasons to transform.}
}
\value{
Numeric year.
}
\description{
Transform season to year. The herring 'season' is a combination of two
fishery years. For example, season '20123' indicates 2012 and 2013. The input
file for the stock assessment requires an actual year which we define as the
second (i.e., later) year, 2013.
}
\seealso{
Other utility: 
\code{\link{mean_na}()},
\code{\link{mean_na_roll}()},
\code{\link{mean_na_weight}()},
\code{\link{sum_na}()}
}
\concept{utility}
