% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_catch}
\alias{siscah_catch}
\title{SISCAH input: catch data.}
\usage{
siscah_catch(
  catch,
  structure = "Section",
  kilo = TRUE,
  sections = undefined_sections,
  n_digits = 3
)
}
\arguments{
\item{catch}{Tibble. Catch data from \code{\link{load_catch}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Section".}

\item{kilo}{Logical. Return values in thousands. Default TRUE.}

\item{sections}{Tibble. Undefined Sections (default
\code{\link{undefined_sections}}).}

\item{n_digits}{Numeric. Number of decimal places for values.}
}
\value{
Tibble with catch data aggregated by Year, Period, and "structure".
Catch is in tonnes for gear 1 (other; reduction, food and bait, as well as
special use), gear 2 (roe seine), and gear 3 (roe gillnet), and in pounds
for gear 4 (spawn on kelp; SOK); values are in thousands if \code{kilo} is
\code{TRUE}.
}
\description{
Make input data for SISCAH: catch data.
}
\examples{
library(magrittr)
library(here)
library(dplyr)
data(undefined_sections)
if (file.exists(here("Examples", "catch_raw.rds"))) {
  catch_raw <- readRDS(file = here("Examples", "catch_raw.rds"))
} else {
  example(load_catch)
}
catch <- catch_raw \%>\%
  filter(Region == "PRD") \%>\%
  siscah_catch(structure = "Section")
}
\seealso{
Other SISCAH: 
\code{\link{siscah_bio}()},
\code{\link{siscah_spawn}()}
}
\concept{SISCAH}
