% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siscah.R
\name{siscah_spawn}
\alias{siscah_spawn}
\title{SISCAH input: spawn data.}
\usage{
siscah_spawn(
  spawn,
  structure = "Section",
  kilo = TRUE,
  sections = undefined_sections,
  n_digits = 3
)
}
\arguments{
\item{spawn}{Tibble. Spawn index data from \code{\link{load_spawn}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Section".}

\item{kilo}{Logical. Return values in thousands. Default TRUE.}

\item{sections}{Tibble. Undefined Sections (default
\code{\link{undefined_sections}}).}

\item{n_digits}{Numeric. Number of decimal places for values.}
}
\value{
Tibble with spawn index aggregated by Year and "structure". Spawn is
in tonnes for surface and dive surveys; values are in thousands if \code{kilo}
is \code{TRUE}.
}
\description{
Make input data for SISCAH: spawn index data.
}
\examples{
library(magrittr)
library(here)
library(dplyr)
data(undefined_sections)
example(load_spawn)
spawn <- spawn_raw \%>\%
  filter(Region == "PRD") \%>\%
  siscah_spawn(structure = "Section")
}
\seealso{
Other SISCAH: 
\code{\link{siscah_bio}()},
\code{\link{siscah_catch}()}
}
\concept{SISCAH}
