% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_nA}
\alias{sum_nA}
\title{Calculate sum.}
\usage{
sum_nA(x, omit_na = TRUE)
}
\arguments{
\item{x}{Numeric. Values to sum.}

\item{omit_na}{Logical. Keep or remove NAs.}
}
\description{
Calculate sum if there are non-NA values, return NA if all values are NA.
}
