% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unbalanced_sampling}
\alias{unbalanced_sampling}
\title{Unbalanced sampling.}
\format{
List of lists with one element for each region indicating the spatial
structure, year(s) to use to fix unbalanced sampling (historic), and
year(s) that were unbalanced.
}
\usage{
unbalanced_sampling
}
\description{
Unbalanced sampling: indicates region, structure, and years where unbalanced
sampling occurred.
}
\examples{
data(unbalanced_sampling)
unbalanced_sampling
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}}
}
\concept{data}
\keyword{datasets}
