% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{undefined_sections}
\alias{undefined_sections}
\title{Undefined Sections.}
\format{
Tibble with Statistical Areas and Sections.
}
\usage{
undefined_sections
}
\description{
Spatially undefined Pacific Herring Sections, typically used for catch in a
known Statistical Area with an unknown Section. These Section codes end in
"0", for example, Section "050" indicates that catch is in Statistical Area
"05". Note that there are two exceptions that are actual areas: Sections
"220" and "280".
}
\examples{
data(undefined_sections)
undefined_sections
}
\seealso{
Other data: 
\code{\link{codes_disposal}},
\code{\link{codes_gear}},
\code{\link{codes_group}},
\code{\link{codes_period}},
\code{\link{codes_source}},
\code{\link{conv_factors}},
\code{\link{database_info}},
\code{\link{unbalanced_sampling}}
}
\concept{data}
\keyword{datasets}
