% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_sections}
\alias{check_sections}
\title{Check for undefined Sections.}
\usage{
check_sections(dat, dat_name, sections = undefined_sections)
}
\arguments{
\item{dat}{Tibble. Data to check.}

\item{dat_name}{Character. Name of data (for warning message). Default
"Catch."}

\item{sections}{Tibble. Undefined Sections (default
\code{\link{undefined_sections}}).}
}
\value{
Warning with list of undefined sections present in data.
}
\description{
Check for data in undefined Sections using \code{\link{undefined_sections}}.
}
\examples{
data(codes_group)
undefined <- check_sections(dat = codes_group, dat_name = "Groups")
}
\concept{check}
