% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_catch}
\alias{load_catch}
\title{Load catch data.}
\usage{
load_catch(
  db_conn,
  db_location,
  area_table,
  year_start = SpawnIndex::pars$years$assess,
  quiet = FALSE
)
}
\arguments{
\item{db_conn}{Character. Herring database connection; see example.}

\item{db_location}{Character. Herring database location; see example.}

\item{area_table}{Tibble. Area information from \code{\link{load_area_data}}.}

\item{year_start}{Integer. First year to include in time series. Default is
from \code{\link{SpawnIndex::pars$years$assess}}.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\value{
Tibble with catch data including spatial and temporal information.
}
\description{
Load and wrangle catch data from the database.
}
\examples{
library(SpawnIndex)
library(DBI)
data(regions)
data(pars)
data(codes_group)
data(conv_factors)
herring_conn <- list(
  driver = "SQL Server Native Client 11.0",
  server = "DFBCV9TWVASP001\\\\SQLEXPRESS16",
  database = "Herring",
  trusted = "Yes"
)
area_loc <- list(
  schema = "Location",
  tables = list(sections = "Sections", locations = "Location"),
  columns = list(
    sections = c("SAR", "Section"),
    locations = c(
      "Loc_Code", "Location", "StatArea", "Section", "Bed",
       "Location_Latitude", "Location_Longitude")
  )
)
areas <- load_area_data(
  db = herring_conn, where = area_loc, reg = "All", groups = codes_group,
  quiet = TRUE
)
catch_loc <- list(
  schema = "Catch",
  tables = list(
    tab_catch = "tCatchData", hail_catch = "HailCatch",
    sok_catch = "SpawnOnKelp"
  ),
  columns = list(
    tab_catch = c(
      "Season", "LocationCode","GearCode", "DisposalCode", "Catch", "Date"
    ),
    hail_catch = c(
      "Active", "Season", "Section", "GearCode", "DisposalCode", "CatchTons"
    ),
    sok_catch = c(
      "Season", "Section", "GearCode", "DisposalCode", "ProductLanded"
    )
  )
)
catch_raw <- load_catch(
  db_conn = herring_conn, db_location = catch_loc, area_table = areas
)
}
\seealso{
Other loaders: 
\code{\link{load_bio}()}
}
\concept{loaders}
