% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_catch}
\alias{load_catch}
\title{Load catch data from the database.}
\usage{
load_catch(db_conn, db_location, area_table, quiet = FALSE)
}
\arguments{
\item{db_conn}{Character. Herring database connection; see example.}

\item{db_location}{Character. Herring database location; see example.}

\item{area_table}{Tibble. Area information from \code{\link{load_area_data}}.}

\item{quiet}{Logical. Suppress messages; default is FALSE.}
}
\description{
Load and wrangle catch data from the database.
}
\examples{
library(SpawnIndex)
library(DBI)
data(regions)
data(pars)
data(codes_group)
data(conv_factors)
herring_conn <- list(
  driver = "SQL Server Native Client 11.0",
  server = "DFBCV9TWVASP001\\\\SQLEXPRESS16",
  database = "Herring",
  trusted = "Yes"
)
area_loc <- list(
  schema = "Location",
  tables = list(sections = "Sections", locations = "Location"),
  columns = list(
    sections = c("SAR", "Section"),
    locations = c(
      "Loc_Code", "Location", "StatArea", "Section", "Bed",
       "Location_Latitude", "Location_Longitude")
  )
)
catch_loc <- list(
  schema = "Catch",
  tables = list(
    tCatch = "tCatchData", hCatch = "HailCatch", sokCatch = "SpawnOnKelp"
  ),
  columns = list(
    tCatch = c(
      "Season", "LocationCode","GearCode", "DisposalCode", "Catch", "Date"
    ),
    hCatch = c(
      "Active", "Season", "Section", "GearCode", "DisposalCode", "CatchTons"
    ),
    sokCatch = c(
      "Season", "Section", "GearCode", "DisposalCode", "ProductLanded"
    )
  )
)
areas <- load_area_data(
  db = herring_conn, where = area_loc, reg = "CC", sec_sub = NULL,
  groups = codes_group
)
catch_raw <- load_catch(
  db_conn = herring_conn, db_location = catch_loc, area_table = areas
)
}
\concept{loaders}
