% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{season_to_year}
\alias{season_to_year}
\title{Transform season to year.}
\usage{
season_to_year(dat)
}
\arguments{
\item{dat}{Numeric. Values to transform.}
}
\description{
Transform season to year. The herring 'season' column is a combination of the
two fishery years: for example, season '20123' indicates the years 2012 and
2013. The input *.dat file for the analysis using ADMB requires this to be an
actual year, which we define as the second (i.e., later) year, 2013. This
function takes in a vector of seasons (dat), and outputs a vector of years
(res).
}
