% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_bio}
\alias{siscah_bio}
\title{SISCAH input: biological data.}
\usage{
siscah_bio(
  bio,
  structure = "Region",
  age_min_number = 1,
  age_min_weight = 2,
  age_max = 10,
  kilo_weight = TRUE,
  year_start = SpawnIndex::pars$years$assess,
  year_end = 2024,
  n_digits = 4
)
}
\arguments{
\item{bio}{Tibble. Biological data from \code{\link{load_bio}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Region".}

\item{age_min_number}{Integer. Youngest age to consider for number-at-age:
omit younger ages. Default 1.}

\item{age_min_weight}{Integer. Youngest age to consider for weight-at-age:
omit younger ages. Default 2.}

\item{age_max}{Integer. Plus year: aggregate ages that are greater than
\code{age_max}. Default 10.}

\item{kilo_weight}{Logical. Return values in thousands for weight. Default
TRUE.}

\item{year_start}{Integer. First year to include in time series. Default is
from \code{SpawnIndex::pars$years$assess}.}

\item{year_end}{Integer. Last year to include in time series. Default is
2024.}

\item{n_digits}{Numeric. Number of decimal places for values.}
}
\value{
List of tibbles with biological data aggregated by Year, Period, and
"structure", including spatial and temporal information. List has three
tibbles: number-at-age and two with weight-at-age in grams.
}
\description{
Make input data for SISCAH: biological data (number- and weight-at-age).
}
\examples{
library(magrittr)
library(here)
library(dplyr)
if(file.exists(here("Examples", "bio_raw.rds"))){
  bio_raw <- readRDS(file = here("Examples", "bio_raw.rds"))
} else{
  example(load_bio)
}
dat <- bio_raw \%>\%
  filter(Region == "PRD") \%>\%
  siscah_bio(structure = "Region")
}
\seealso{
Other SISCAH: 
\code{\link{siscah_catch}()},
\code{\link{siscah_spawn}()}
}
\concept{SISCAH}
