% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_bio}
\alias{siscah_bio}
\title{SISCAH input: biological data.}
\usage{
siscah_bio(bio, structure = "Region", age_min = 1, age_max = 10, n_digits = 3)
}
\arguments{
\item{bio}{Tibble. Biological data from \code{\link{load_bio}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Region".}

\item{age_min}{Integer. Youngest age to consider: omit younger ages. Default
1.}

\item{age_max}{Integer. Plus year: aggregate ages that are greater than
\code{age_max}. Default 10.}

\item{n_digits}{Numeric. Number of decimal places for values. Default 3.}
}
\value{
List of tibbles with biological data aggregated by Year, Period, and
"structure", including spatial and temporal information. List has two
tibbles: number-at-age and weight-at-age in grams.
}
\description{
Make input data for SISCAH: biological data (number- and weight-at-age).
}
\examples{
\dontrun{
bio_raw <- readRDS(file = here("Examples", "bio_raw.rds"))
}
example(load_bio)
dat <- bio_raw \%>\%
  filter(Region == "PRD") \%>\%
  siscah_bio(structure = "StatArea")
}
\seealso{
Other SISCAH: 
\code{\link{siscah_catch}()},
\code{\link{siscah_spawn}()}
}
\concept{SISCAH}
