% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_catch}
\alias{siscah_catch}
\title{Input file for SISCAH: catch data.}
\usage{
siscah_catch(catch, structure = "Region", file)
}
\arguments{
\item{catch}{Tibble. Catch data from \code{\link{load_catch}}.}

\item{structure}{Character. Column name indicating stock structure; must be
present in \code{catch}. Default is "Region".}

\item{file}{Character. Output path and file name.}
}
\description{
Make the input file for SISCAH: catch data.
}
\concept{SISCAH}
