% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_catch}
\alias{siscah_catch}
\title{SISCAH input: catch data.}
\usage{
siscah_catch(catch, structure = "Region", kilo = TRUE, n_digits = 3)
}
\arguments{
\item{catch}{Tibble. Catch data from \code{\link{load_catch}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Region".}

\item{kilo}{Logical. Return values in thousands. Default TRUE.}

\item{n_digits}{Numeric. Number of decimal places for values.}
}
\value{
Tibble with catch data aggregated by Year, Period, and "structure",
including spatial and temporal information.
}
\description{
Make input data for SISCAH: catch data.
}
\examples{
library(magrittr)
library(here)
library(dplyr)
if(file.exists(here("Examples", "catch_raw.rds"))){
  catch_raw <- readRDS(file = here("Examples", "catch_raw.rds"))
} else{
  example(load_catch)
}
dat <- catch_raw \%>\%
  filter(Region == "PRD") \%>\%
  siscah_catch(structure = "Region")
}
\seealso{
Other SISCAH: 
\code{\link{siscah_bio}()},
\code{\link{siscah_spawn}()}
}
\concept{SISCAH}
