% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_catch}
\alias{siscah_catch}
\title{Input file for SISCAH: catch data.}
\usage{
siscah_catch(catch, structure = Region, kilo = TRUE, n_digits = 3)
}
\arguments{
\item{catch}{Tibble. Catch data from \code{\link{load_catch}}.}

\item{structure}{Character. Column name indicating stock structure. Use
backticks. Default is \code{Region}.}

\item{kilo}{Logical. Return values in thousands. Default TRUE.}

\item{n_digits}{Numeric. Number of decimal places for values. Default 3.}
}
\value{
Tibble with catch data in thousands aggregated by Year, Period, and
\code{structure}, including spatial and temporal information.
}
\description{
Make the input file for SISCAH: catch data.
}
\seealso{
Other SISCAH: 
\code{\link{siscah_spawn}()}
}
\concept{SISCAH}
