% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{siscah_spawn}
\alias{siscah_spawn}
\title{SISCAH input: spawn data.}
\usage{
siscah_spawn(spawn, structure = "Region", kilo = TRUE, n_digits = 3)
}
\arguments{
\item{spawn}{Tibble. Spawn index data from \code{\link{load_spawn}}.}

\item{structure}{Character. Column name indicating stock structure. Default
is "Region".}

\item{kilo}{Logical. Return values in thousands. Default TRUE.}

\item{n_digits}{Numeric. Number of decimal places for values. Default 3.}
}
\value{
Tibble with catch data in thousands aggregated by Year, Period, and
"structure", including spatial and temporal information.
}
\description{
Make input data for SISCAH: spawn index data.
}
\seealso{
Other SISCAH: 
\code{\link{siscah_bio}()},
\code{\link{siscah_catch}()}
}
\concept{SISCAH}
